+++ client/web/src/enterprise/insights/components/insights-view-grid/components/insight-context-menu/InsightContextMenu.tsx
@@ -1,6 +1,6 @@
 import { Menu, MenuButton, MenuItem, MenuItems, MenuLink, MenuPopover } from '@reach/menu-button'
 import classnames from 'classnames'
-import CheckIcon from 'mdi-react/CheckIcon'
+import { noop } from 'lodash'
 import DotsVerticalIcon from 'mdi-react/DotsVerticalIcon'
 import React, { useContext } from 'react'
 import { Link } from 'react-router-dom'
@@ -40,6 +40,7 @@ export const InsightContextMenu: React.FunctionComponent<InsightCardMenuProps> =
                     <MenuButton
                         data-testid="InsightContextMenuButton"
                         className={classnames(menuButtonClassName, 'btn btn-outline p-1', styles.button)}
+                        aria-label="Insight options"
                     >
                         <DotsVerticalIcon
                             className={classnames(styles.buttonIcon, { [styles.buttonIconActive]: isOpen })}
@@ -54,27 +55,38 @@ export const InsightContextMenu: React.FunctionComponent<InsightCardMenuProps> =
                             <MenuLink
                                 as={Link}
                                 data-testid="InsightContextMenuEditLink"
-                                className={classnames('btn btn-outline', styles.item)}
+                                className={styles.item}
                                 to={editUrl}
                             >
                                 Edit
                             </MenuLink>
 
                             {showYAxisToggleMenu && (
-                                <MenuLink
+                                <MenuItem
+                                    role="menuitemcheckbox"
                                     data-testid="InsightContextMenuEditLink"
-                                    className={classnames('btn btn-outline border-bottom', styles.item)}
-                                    onClick={onToggleZeroYAxisMin}
+                                    className={classnames(
+                                        'd-flex align-items-center justify-content-between',
+                                        styles.item
+                                    )}
+                                    onSelect={onToggleZeroYAxisMin}
+                                    aria-checked={zeroYAxisMin}
                                 >
-                                    <CheckIcon size={16} className={classnames('mr-2', { 'd-none': !zeroYAxisMin })} />{' '}
-                                    Start Y Axis at 0
-                                </MenuLink>
+                                    <input
+                                        type="checkbox"
+                                        aria-hidden="true"
+                                        checked={zeroYAxisMin}
+                                        onChange={noop}
+                                        tabIndex={-1}
+                                    />
+                                    <span>Start Y Axis at 0</span>
+                                </MenuItem>
                             )}
 
                             <MenuItem
                                 data-testid="insight-context-menu-delete-button"
                                 onSelect={() => onDelete(insightID)}
-                                className={classnames('btn btn-outline-', styles.item)}
+                                className={styles.item}
                             >
                                 Delete
                             </MenuItem>