+++ client/web/src/search/home/SearchPage.story.tsx
@@ -4,16 +4,16 @@ import { createMemoryHistory } from 'history'
 import React from 'react'
 
 import { NOOP_TELEMETRY_SERVICE } from '@sourcegraph/shared/src/telemetry/telemetryService'
+import {
+    mockFetchAutoDefinedSearchContexts,
+    mockFetchSearchContexts,
+    mockGetUserSearchContextNamespaces,
+} from '@sourcegraph/shared/src/testing/searchContexts/testHelpers'
 import { ThemeProps } from '@sourcegraph/shared/src/theme'
 import { extensionsController } from '@sourcegraph/shared/src/util/searchTestHelpers'
 
 import { WebStory } from '../../components/WebStory'
 import { SearchPatternType } from '../../graphql-operations'
-import {
-    mockFetchAutoDefinedSearchContexts,
-    mockFetchSearchContexts,
-    mockGetUserSearchContextNamespaces,
-} from '../../searchContexts/testHelpers'
 import { ThemePreference } from '../../theme'
 import { _fetchRecentFileViews, _fetchRecentSearches, _fetchSavedSearches, authUser } from '../panels/utils'
 
@@ -44,12 +44,12 @@ const defaultProps = (props: ThemeProps): SearchPageProps => ({
     platformContext: {} as any,
     keyboardShortcuts: [],
     versionContext: undefined,
+    searchContextsEnabled: true,
     showSearchContext: false,
     showSearchContextManagement: false,
     selectedSearchContextSpec: '',
     setSelectedSearchContextSpec: () => {},
     defaultSearchContextSpec: '',
-    showRepogroupHomepage: false,
     showEnterpriseHomePanels: false,
     showOnboardingTour: false,
     showQueryBuilder: false,
@@ -83,13 +83,9 @@ add('Cloud with panels', () => (
     </WebStory>
 ))
 
-add('Cloud without repogroups or panels', () => (
-    <WebStory>{webProps => <SearchPage {...defaultProps(webProps)} isSourcegraphDotCom={true} />}</WebStory>
-))
-
-add('Cloud with repogroups', () => (
+add('Cloud with community search contexts', () => (
     <WebStory>
-        {webProps => <SearchPage {...defaultProps(webProps)} isSourcegraphDotCom={true} showRepogroupHomepage={true} />}
+        {webProps => <SearchPage {...defaultProps(webProps)} isSourcegraphDotCom={true} authenticatedUser={null} />}
     </WebStory>
 ))
 
@@ -99,7 +95,7 @@ add('Cloud with notebook onboarding', () => (
             <SearchPage
                 {...defaultProps(webProps)}
                 isSourcegraphDotCom={true}
-                showRepogroupHomepage={true}
+                authenticatedUser={null}
                 featureFlags={new Map([['search-notebook-onboarding', true]])}
             />
         )}