+++ dev/ci/test/qa/test.sh
@@ -4,6 +4,7 @@
 source /root/.profile
 root_dir="$(dirname "${BASH_SOURCE[0]}")/../../../.."
 cd "$root_dir"
+root_dir=$(pwd)
 
 set -ex
 
@@ -21,9 +22,11 @@ trap cleanup EXIT
 CONTAINER=sourcegraph-server
 
 docker_logs() {
+  pushd "$root_dir"
   LOGFILE=$(docker inspect ${CONTAINER} --format '{{.LogPath}}')
   cp "$LOGFILE" $CONTAINER.log
   chmod 744 $CONTAINER.log
+  popd
 }
 
 if [[ $VAGRANT_RUN_ENV = "CI" ]]; then
@@ -50,6 +53,8 @@ set -x
 echo "TEST: Checking Sourcegraph instance is accessible"
 curl -f http://localhost:7080
 curl -f http://localhost:7080/healthz
+echo "TEST: Downloading Puppeteer"
+yarn --cwd client/shared run download-puppeteer-browser
 echo "TEST: Running tests"
 # Run all tests, and error if one fails
 test_status=0