+++ dev/ci/test/upgrade/test.sh
@@ -4,6 +4,7 @@
 source /root/.profile
 root_dir="$(dirname "${BASH_SOURCE[0]}")/../../../.."
 cd "$root_dir"
+root_dir=$(pwd)
 
 set -ex
 
@@ -65,9 +66,11 @@ fi
 # Upgrade to current candidate image. Capture logs for the attempted upgrade.
 CONTAINER=sourcegraph-new
 docker_logs() {
+  pushd "$root_dir"
   LOGFILE=$(docker inspect ${CONTAINER} --format '{{.LogPath}}')
   cp "$LOGFILE" $CONTAINER.log
   chmod 744 $CONTAINER.log
+  popd
 }
 IMAGE=us.gcr.io/sourcegraph-dev/server:$CANDIDATE_VERSION CLEAN="false" ./dev/run-server-image.sh -d --name $CONTAINER
 trap docker_logs exit
@@ -78,6 +81,8 @@ echo "TEST: Checking Sourcegraph instance is accessible"
 curl -f http://localhost:7080
 curl -f http://localhost:7080/healthz
 echo "TEST: Running tests"
+echo "TEST: Downloading Puppeteer"
+yarn --cwd client/shared run download-puppeteer-browser
 pushd client/web
 yarn run test:regression:core
 popd