+++ internal/repos/metrics.go
@@ -149,16 +149,9 @@ AND deleted_at IS NULL
 		count, err := scanCount(`
 -- source: internal/repos/metrics.go:src_repoupdater_user_repos_total
 SELECT COUNT(*)
-FROM external_service_repos esr
-JOIN external_services es ON (
-  es.id = esr.external_service_id AND
-  es.namespace_user_id IS NOT NULL AND
-  es.deleted_at IS NULL
-)
-JOIN repo ON (
-  repo.id = esr.repo_id AND
-  repo.deleted_at IS NULL
-)`)
+FROM external_service_repos
+WHERE user_id IS NOT NULL
+`)
 		if err != nil {
 			log15.Error("Failed to get total user repositories", "err", err)
 			return 0